# [Enhancement] add scan op dop multiplier