--- a/be/src/exec/vectorized/connector_scan_node.cpp
+++ b/be/src/exec/vectorized/connector_scan_node.cpp
@@ -560,4 +560,8 @@ int ConnectorScanNode::io_tasks_per_scan_operator() const {
     return config::connector_io_tasks_per_scan_operator;
 }
 
+int ConnectorScanNode::scan_operator_dop_multiplier() const {
+    return config::connector_scan_operator_dop_multiplier;
+}
+
 } // namespace starrocks::vectorized